apis/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.1
starlette-context==0.4.0
pydantic-settings==2.9.1
loguru==0.7.3
kubernetes==31.0.0
tritonclient[http]==2.56.0
opencv-python==4.11.0.86
numpy==1.26.4
prometheus-client==0.21.1
opentelemetry-sdk==3.21.0
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-instrumentation-logging==0.53b1